[Fix] #448 - TagTypes 분리
