+++ SOPT-iOS/Projects/Features/HomeFeature/Sources/HomeScene/Cells/Group/GroupCardCVC.swift
@@ -11,45 +11,6 @@ import UIKit
 import Core
 import DSKit
 
-enum RecruitmentStatusType: String {
-    case beforeStart = "BEFORE_START"
-    case applyAble = "APPLY_ABLE"
-    case recruitmentComplete = "RECRUITMENT_COMPLETE"
-    
-    var text: String {
-        switch self {
-        case .beforeStart:
-            return I18N.Home.Group.beforeStart
-        case .applyAble:
-            return I18N.Home.Group.applyAble
-        case .recruitmentComplete:
-            return I18N.Home.Group.recruitmentComplete
-        }
-    }
-    
-    var textColor: UIColor {
-        switch self {
-        case .beforeStart:
-            return DSKitAsset.Colors.gray800.color
-        case .applyAble:
-            return DSKitAsset.Colors.gray800.color
-        case .recruitmentComplete:
-            return DSKitAsset.Colors.gray800.color
-        }
-    }
-
-    var backgroundColor: UIColor {
-        switch self {
-        case .beforeStart:
-            return DSKitAsset.Colors.attention.color
-        case .applyAble:
-            return DSKitAsset.Colors.secondary.color
-        case .recruitmentComplete:
-            return DSKitAsset.Colors.gray100.color
-        }
-    }
-}
-
 final class GroupCardCVC: UICollectionViewCell {
     
     // MARK: - Properties
@@ -192,7 +153,7 @@ extension GroupCardCVC {
     }
     
     /// 카테고리 + 모임글 타이틀
-    private func makeTitleAttributedString(category: GroupCategoryType, title: String) {
+    private func makeTitleAttributedString(category: GroupCategoryTagType, title: String) {
         let attributedText = "\(category.text) \(title)"
         self.titleLabel.text = attributedText
         self.titleLabel.partColorChange(targetString: category.text, textColor: category.textColor)
