+++ SOPT-iOS/Projects/Features/HomeFeature/Sources/HomeScene/ViewModel/HomeForMemberViewModel.swift
@@ -38,35 +38,12 @@ struct InsightInfo {
 
 struct GroupInfo {
     let title: String
-    let category: GroupCategoryType
+    let category: GroupCategoryTagType
     let canJoinOnlyActiveGeneration: Bool
     let joinableParts: [String]
     let canJoinAllParts: Bool
-    let status: RecruitmentStatusType
+    let status: RecruitmentStatusTagType
     let imageURL: String
-}
-
-enum GroupCategoryType: String {
-    case event = "EVENT"
-    case study = "STUDY"
-    
-    var text: String {
-        switch self {
-        case .event:
-            return "행사"
-        case .study:
-            return "스터디"
-        }
-    }
-    
-    var textColor: UIColor {
-        switch self {
-        case .event:
-            return DSKitAsset.Colors.success.color
-        case .study:
-            return DSKitAsset.Colors.secondary.color
-        }
-    }
 }
 
 public class HomeForMemberViewModel: HomeForMemberViewModelType {
